Rename(HttpResponse) : Http response 클래스명 변경
HttpRequest 클래스와 통일감을 주기 위해 Response -> HttpResponse로 클래스명을 변경했다
